Add skopeo internal package
