+++ cmd/getsizeworker/main.go
@@ -6,7 +6,6 @@ import (
 	"fmt"
 	"log"
 	"os"
-	"os/exec"
 	"path/filepath"
 	"strings"
 	"sync"
@@ -19,6 +18,7 @@ import (
 	"github.com/vpereira/trivy_runner/internal/metrics"
 	"github.com/vpereira/trivy_runner/internal/redisutil"
 	"github.com/vpereira/trivy_runner/internal/sentry"
+	"github.com/vpereira/trivy_runner/internal/skopeo"
 	"github.com/vpereira/trivy_runner/internal/util"
 	"github.com/vpereira/trivy_runner/pkg/exec_command"
 	"go.uber.org/zap"
@@ -113,7 +113,7 @@ func main() {
 }
 
 func downloadImageAndGetSize(image, architecture, filePath string) (int64, error) {
-	cmdArgs := GenerateSkopeoCmdArgs(image, filePath, architecture)
+	cmdArgs := skopeo.GenerateSkopeoCmdArgs(image, filePath, architecture)
 
 	logger.Info("Executing skopeo with arguments", zap.String("arguments", strings.Join(cmdArgs, " ")))
 
@@ -164,7 +164,7 @@ func processQueue() {
 	logger.Info("Target directory: ", zap.String("targetDir", targetDir))
 
 	// Get the supported architectures for the image
-	architectures, err := getSupportedArchitectures(imageName)
+	architectures, err := skopeo.GetSupportedArchitectures(imageName)
 
 	if err != nil {
 		errorHandler.Handle(err)
@@ -229,44 +229,6 @@ func processQueue() {
 	prometheusMetrics.IncOpsProcessed()
 }
 
-// GenerateSkopeoInspectCmdArgs generates the command line arguments for the skopeo inspect to fetch all supported architectures
-func GenerateSkopeoInspectCmdArgs(imageName string) []string {
-	return []string{"inspect", "--raw", fmt.Sprintf("docker://%s", imageName)}
-}
-
-// getSupportedArchitectures gets the list of supported architectures for a Docker image.
-func getSupportedArchitectures(image string) ([]string, error) {
-	cmdArgs := GenerateSkopeoInspectCmdArgs(image)
-	cmd := exec.Command("skopeo", cmdArgs...)
-	output, err := cmd.CombinedOutput()
-	if err != nil {
-		return nil, fmt.Errorf("skopeo output: %s, error: %s", string(output), err.Error())
-	}
-
-	var manifest struct {
-		Manifests []struct {
-			Platform struct {
-				Architecture string `json:"architecture"`
-			} `json:"platform"`
-		} `json:"manifests"`
-	}
-	if err := json.Unmarshal(output, &manifest); err != nil {
-		return nil, err
-	}
-
-	var architectures []string
-	for _, m := range manifest.Manifests {
-		architectures = append(architectures, m.Platform.Architecture)
-	}
-
-	// Ensure at least "amd64" is included if no architectures were found
-	if len(architectures) == 0 {
-		architectures = []string{"amd64"}
-	}
-
-	return architectures, nil
-}
-
 // getFileSize returns the size of the file at the given path in bytes.
 func getFileSize(filePath string) (int64, error) {
 	fileInfo, err := os.Stat(filePath)
@@ -274,28 +236,6 @@ func getFileSize(filePath string) (int64, error) {
 		return 0, err
 	}
 	return fileInfo.Size(), nil
-}
-
-func GenerateSkopeoCmdArgs(imageName, targetFilename, architecture string) []string {
-	cmdArgs := []string{"copy", "--remove-signatures"}
-
-	// Check and add registry credentials if they are set
-	registryUsername, usernameSet := os.LookupEnv("REGISTRY_USERNAME")
-	registryPassword, passwordSet := os.LookupEnv("REGISTRY_PASSWORD")
-
-	if usernameSet && passwordSet {
-		cmdArgs = append(cmdArgs, "--src-username", registryUsername, "--src-password", registryPassword)
-	}
-
-	// Add architecture override if specified
-	if architecture != "" {
-		cmdArgs = append(cmdArgs, "--override-arch", architecture)
-	}
-
-	// Add the rest of the command source image and destination tar
-	cmdArgs = append(cmdArgs, fmt.Sprintf("docker://%s", imageName), fmt.Sprintf("docker-archive://%s", targetFilename))
-
-	return cmdArgs
 }
 
 // sanitizeImageName replaces slashes and colons in the image name with underscores.
