+++ cmd/getsizeworker/main_test.go
@@ -2,7 +2,6 @@ package main
 
 import (
 	"os"
-	"reflect"
 	"testing"
 
 	"github.com/alicebob/miniredis/v2"
@@ -57,65 +56,4 @@ func TestProcessQueue(t *testing.T) {
 	defer prometheus.Unregister(commandExecutionHistogram)
 
 	processQueue()
-}
-
-func TestGenerateSkopeoCmdArgs(t *testing.T) {
-	// Define test cases
-	tests := []struct {
-		name           string
-		imageName      string
-		architecture   string
-		targetDir      string
-		envUsername    string
-		envPassword    string
-		expectedResult []string
-	}{
-		{
-			name:         "without credentials",
-			imageName:    "registry.example.com/myimage:latest",
-			targetDir:    "/tmp/targetdir",
-			architecture: "amd64",
-			expectedResult: []string{
-				"copy", "--remove-signatures",
-				"--override-arch", "amd64",
-				"docker://registry.example.com/myimage:latest",
-				"docker-archive:///tmp/targetdir",
-			},
-		},
-		{
-			name:         "with credentials",
-			imageName:    "registry.example.com/myimage:latest",
-			targetDir:    "/tmp/targetdir",
-			envUsername:  "testuser",
-			envPassword:  "testpass",
-			architecture: "amd64",
-			expectedResult: []string{
-				"copy", "--remove-signatures",
-				"--src-username", "testuser", "--src-password", "testpass",
-				"--override-arch", "amd64",
-				"docker://registry.example.com/myimage:latest",
-				"docker-archive:///tmp/targetdir",
-			},
-		},
-	}
-
-	for _, tc := range tests {
-		t.Run(tc.name, func(t *testing.T) {
-			// Set up environment variables if needed
-			if tc.envUsername != "" && tc.envPassword != "" {
-				os.Setenv("REGISTRY_USERNAME", tc.envUsername)
-				os.Setenv("REGISTRY_PASSWORD", tc.envPassword)
-				defer os.Unsetenv("REGISTRY_USERNAME")
-				defer os.Unsetenv("REGISTRY_PASSWORD")
-			}
-
-			// Call the method under test
-			result := GenerateSkopeoCmdArgs(tc.imageName, tc.targetDir, tc.architecture)
-
-			// Verify the result
-			if !reflect.DeepEqual(result, tc.expectedResult) {
-				t.Errorf("GenerateSkopeoCmdArgs(%s, %s) got %v, want %v", tc.imageName, tc.targetDir, result, tc.expectedResult)
-			}
-		})
-	}
 }
